\name{nsclc}
\alias{A549_1}
\alias{A549_2}
\alias{Calu1_1}
\alias{Calu1_2}
\alias{H322_1}
\alias{H322_2}
\alias{HCC2429_1}
\alias{HCC2429_2}
\docType{data} 
\title{Results from a compound screen on 84 NSCLC cell lines.}
\description{
  For purpose of computational, lesion-based prediction of compound
  activity in 84 non-small cell lung cancer (NSCLC) cell lines, these
  were treated with a selection of compounds to measure the respective
  concentration where 50 per cent of cell growth was inhibited after a
  predefined time period. Most of the screening experiments were
  performed on 384 (16 x 24)-well plates, were the experimental setup was
  designed according to the arrangement in the files
  \code{"mpi384_measure.txt"}, \code{"mpi384_control.txt"} and
  \code{"mpi384_dilution.txt"} that are installed together with the
  package.

  These data sets are the direct output from a signal reader of the
  384-well plates for the A549, Calu1, H322 and H2429 cancer cell lines
  selected from the 84 NSCLC cell lines collection. A tab-delimited version
  is installed together with the package.
}
\references{
  Frommolt P, Thomas RK (2008): Standardized high-throughput evaluation
  of cell-based compound screens. BMC Bioinformatics, 9(1): 475

  Sos ML, Michel K, Zander T, Weiss J, Frommolt P, et al. (2009): Predicting drug susceptibility
  in non-small cell lung cancers based on genetic lesions. J Clin
  Invest, 119(6): 1727-40
}
\keyword{datasets}
